Suppression du lien entre lot et programme
