+++ conventions/management/commands/deduplicate_programme.py
@@ -140,10 +140,6 @@ class Command(BaseCommand):
                 programmes[0].anru = bool(programmes[0].anru)
                 programmes[0].save()
                 for p in programmes[1:]:
-                    # FIXME : A supprimer quand lots n'a plus programme_id
-                    for lot in p.lots.all():
-                        lot.programme = programmes[0]
-                        lot.save()
                     for convention in p.conventions.all():
                         convention.programme = programmes[0]
                         convention.save()
