+++ conventions/tests/custom_filters/test_main_custom_filters.py
@@ -676,61 +676,47 @@ class CustomFiltersTest(TestCase):
         self.assertTrue(custom_filters.display_create_avenant(self.convention))
 
         cloned_programme = self.convention.programme.clone()
-        cloned_lot = self.convention.lot.clone(cloned_programme)
         av = Convention.objects.create(
             statut=ConventionStatut.A_SIGNER.label,
             parent=self.convention,
             programme=cloned_programme,
         )
-        av.save()
-
-        cloned_lot.convention = av
-        cloned_lot.save()
+        self.convention.lot.clone(convention=av)
 
         self.assertTrue(custom_filters.display_create_avenant(self.convention))
         av.delete()
 
         cloned_programme = self.convention.programme.clone()
-        cloned_lot = self.convention.lot.clone(cloned_programme)
         av = Convention.objects.create(
             statut=ConventionStatut.SIGNEE.label,
             parent=self.convention,
             programme=cloned_programme,
         )
-        av.save()
-
-        cloned_lot.convention = av
-        cloned_lot.save()
+        self.convention.lot.clone(convention=av)
 
         self.assertTrue(custom_filters.display_create_avenant(self.convention))
         av.delete()
 
         cloned_programme = self.convention.programme.clone()
-        cloned_lot = self.convention.lot.clone(cloned_programme)
 
         av = Convention.objects.create(
             statut=ConventionStatut.RESILIEE.label,
             parent=self.convention,
             programme=cloned_programme,
         )
-        av.save()
-
-        cloned_lot.convention = av
-        cloned_lot.save()
+        self.convention.lot.clone(convention=av)
 
         self.assertTrue(custom_filters.display_create_avenant(self.convention))
         av.delete()
 
         cloned_programme = self.convention.programme.clone()
-        cloned_lot = self.convention.lot.clone(cloned_programme)
         ongoing_avenant = Convention.objects.create(
             statut=ConventionStatut.PROJET.label,
             parent=self.convention,
             programme=cloned_programme,
         )
-        cloned_lot.convention = ongoing_avenant
-        cloned_lot.save()
-        ongoing_avenant.save()
+        self.convention.lot.clone(convention=ongoing_avenant)
+
         ongoing_avenant = Convention.objects.get(pk=ongoing_avenant.pk)
         self.assertFalse(custom_filters.display_create_avenant(self.convention))
         ongoing_avenant.statut = ConventionStatut.INSTRUCTION.label
@@ -810,16 +796,13 @@ class CustomFiltersTest(TestCase):
                         )
 
         cloned_programme = self.convention.programme.clone()
-        cloned_lot = self.convention.lot.clone(cloned_programme)
 
         avenant = Convention.objects.create(
             statut=ConventionStatut.SIGNEE.label,
             parent=self.convention,
             programme=cloned_programme,
         )
-
-        cloned_lot.convention = avenant
-        cloned_lot.save()
+        self.convention.lot.clone(convention=avenant)
 
         avenant.type1and2 = None
         avenant.programme.bailleur.sous_nature_bailleur = (
@@ -914,16 +897,13 @@ class CustomFiltersTest(TestCase):
                     )
 
         cloned_programme = self.convention.programme.clone()
-        cloned_lot = self.convention.lot.clone(cloned_programme)
 
         avenant = Convention.objects.create(
             statut=ConventionStatut.SIGNEE.label,
             parent=self.convention,
             programme=cloned_programme,
         )
-
-        cloned_lot.convention = avenant
-        cloned_lot.save()
+        self.convention.lot.clone(convention=avenant)
 
         avenant.type1and2 = None
         avenant.programme.bailleur.sous_nature_bailleur = (
