+++ conventions/tests/views/test_preview_view.py
@@ -8,7 +8,7 @@ from django.urls import reverse
 from pytest_django.asserts import assertTemplateUsed
 
 from conventions.models.choices import ConventionStatut
-from conventions.tests.factories import ConventionFactory
+from core.tests.factories import ConventionFactory
 from users.tests.factories import GroupFactory, RoleFactory, UserFactory
 from users.type_models import TypeRole
